backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
python-dotenv==1.2.1
pytest==8.3.4
httpx==0.28.1
redis==5.2.1
fakeredis==2.26.2
